+++ test/packages/gh15/index.ts
@@ -1,0 +1,2 @@
+import f1 from "./file";
+export type F = typeof f1;